requirements.txt
feedparser
python-dotenv
requests
tldextract
cloudscraper